Call client methods only if client is not None.

+++ src/nti/app/products/courseware_scorm/subscribers.py
@@ -33,7 +33,8 @@ def _enrollment_record_created(record, unused_event):
     course = record.CourseInstance
     if ISCORMCourseInstance.providedBy(course):
         client = component.queryUtility(ISCORMCloudClient)
-        client.sync_enrollment_record(record, course)
+        if client is not None:
+            client.sync_enrollment_record(record, course)
 
 
 @component.adapter(ICourseInstanceEnrollmentRecord, IBeforeIdRemovedEvent)
@@ -41,7 +42,8 @@ def _enrollment_record_dropped(record, unused_event):
     course = record.CourseInstance
     if ISCORMCourseInstance.providedBy(course):
         client = component.queryUtility(ISCORMCloudClient)
-        client.delete_enrollment_record(record)
+        if client is not None:
+            client.delete_enrollment_record(record)
 
 
 @component.adapter(IAllCoursesCollection)
